move tag files to a single file
