+++ src/Me/Modals/model.tsx
@@ -1,6 +1,6 @@
 import ISO6391 from 'iso-639-1';
 import countries from 'svg-country-flags/countries.json';
-import tags from './tags';
+import tags from '../../utils/tags.js';
 
 type Option = {
   label: string;
